fixed concurrency issue with summary when listening on multiple servers
