+++ src/engine/commands/start.command.js
@@ -1,11 +1,11 @@
 const { AppConfig } = require("../../../app.config");
+const { FlushCommand } = require("./flush.command");
 const { Command, HelpTip } = require('../../models/command');
 const { Timestamp } = require("../../models/timestamp");
 const { LiteralConstants } = require('../../utils/literal.constants');
 const { Logger } = require('../../utils/logger');
 const { formatTime } = require('../../utils/time.utils');
 const oneline = require('oneline');
-const { Guild } = require("discord.js");
 
 /**
  * 
@@ -61,7 +61,7 @@ function command(appConfig){
             },
             // on live end
             async (live) => {
-                generateTimestampSummary(appConfig, configKey, guild, logger);
+                await FlushCommand(appConfig).callback(message, args, override);
             },
             // on open
             async () => {
@@ -85,7 +85,7 @@ function command(appConfig){
             // flush if stream is offline when start up (IE, bot went down and we missed the onLiveEnd event)
             const liveStatus = await listener.getLiveStatus()
             if(!liveStatus.isLive()){
-                generateTimestampSummary(appConfig, configKey, guild, logger);
+                await FlushCommand(appConfig).callback(message, args, override);
             }
         }, 
         [
@@ -96,59 +96,6 @@ function command(appConfig){
             ),
         ]
     );
-}
-
-/**
- * 
- * @param {AppConfig} appConfig 
- * @param {*} configKey 
- * @param {Guild} guild 
- * @param {console} logger 
- */
- async function generateTimestampSummary(appConfig, configKey, guild, logger){
-    logger.log(`Generating stream-end summary...`);
-    const timestamps = appConfig.TIMESTAMP_STORAGE.getAllTimestamps(guild);
-    if(timestamps && timestamps.length > 0){
-        // get all language channels
-        const channels = [...Object.entries(appConfig.CONFIG_STORAGE.getProperty(configKey, 'output').chat)];
-        for(entry of timestamps){
-            // if timestamp log has content
-            if(entry.length > 1){
-                const language = entry[0];
-                let summary = "";
-                // find corresponding output channel
-                const channel = appConfig.DISCORD_HELPERS.getChannel(guild, channels.find(c => c[0] == language)[1]);
-                for(timestamp of entry[1]){
-                    // check each stored timestamp for upvotes
-                    if(timestamp.length > 1){
-                        const timestampId = timestamp[0];
-                        try{
-                            const timestampMessage = await channel.messages.fetch(timestampId, true, true);
-                            if(timestampMessage){
-                                const upvotes = timestampMessage.reactions.cache.get(LiteralConstants.REACT_UPVOTE_EMOJI);
-                                const upvoteCount = upvotes ? upvotes.count : 0;
-                                const downvotes = timestampMessage.reactions.cache.get(LiteralConstants.REACT_DOWNVOTE_EMOJI);
-                                const downvoteCount = downvotes ? downvotes.count : 0;
-                                if(upvoteCount >= downvoteCount){
-                                    // write to summary log if upvoted
-                                    summary = summary.concat(timestamp[1], '\n');
-                                }
-                            }
-                        }catch(e){
-                            logger.error(`Unable to find message with ID: ${timestampId}: ${e}`);
-                        }
-                    }
-                }
-                if(summary.length > 0){
-                    logger.log(`Posting ${language} summary:\n${summary}`);
-                    const attachment = appConfig.DISCORD_HELPERS.generateAttachment(summary, `${language}-summary.txt`);
-                    channel.send(`The stream has ended. Here's a summary:`, attachment);
-                }
-            }
-        }
-        // flush
-        appConfig.TIMESTAMP_STORAGE.deleteGuildTimestamps(guild);
-    }
 }
 
 module.exports.StartCommand = command;
